Removing utils.go file
