+++ internal/resources/acceptance_tests/cluster_acceptance_test.go
@@ -2,8 +2,11 @@ package acceptance_tests
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"log"
 	"net/http"
+	"os"
 	"regexp"
 	"terraform-provider-capella/internal/api"
 	"testing"
@@ -1537,5 +1540,95 @@ func generateClusterImportIdForResource(resourceReference string) resource.Impor
 		}
 		fmt.Printf("raw state %s", rawState)
 		return fmt.Sprintf("id=%s,organization_id=%s,project_id=%s", rawState["id"], rawState["organization_id"], rawState["project_id"]), nil
+	}
+}
+
+func testAccDeleteCluster(clusterResourceReference, projectResourceReference string) resource.TestCheckFunc {
+	log.Println("Deleting the cluster")
+	return func(s *terraform.State) error {
+		var clusterState, projectState map[string]string
+		for _, m := range s.Modules {
+			if len(m.Resources) > 0 {
+				if v, ok := m.Resources[clusterResourceReference]; ok {
+					clusterState = v.Primary.Attributes
+				}
+				if v, ok := m.Resources[projectResourceReference]; ok {
+					projectState = v.Primary.Attributes
+				}
+			}
+		}
+		data, err := acctest.TestClient()
+		if err != nil {
+			return err
+		}
+		host := os.Getenv("TF_VAR_host")
+		orgid := os.Getenv("TF_VAR_organization_id")
+		authToken := os.Getenv("TF_VAR_auth_token")
+		url := fmt.Sprintf("%s/v4/organizations/%s/projects/%s/clusters/%s", host, orgid, projectState["id"], clusterState["id"])
+		cfg := api.EndpointCfg{Url: url, Method: http.MethodDelete, SuccessStatus: http.StatusAccepted}
+		_, err = data.Client.Execute(
+			cfg,
+			nil,
+			authToken,
+			nil,
+		)
+		if err != nil {
+			return err
+		}
+		return nil
+	}
+}
+
+// retrieveClusterFromServer checks cluster exists in server.
+func retrieveClusterFromServer(data *providerschema.Data, organizationId, projectId, clusterId string) (*clusterapi.GetClusterResponse, error) {
+	url := fmt.Sprintf("%s/v4/organizations/%s/projects/%s/clusters/%s", data.HostURL, organizationId, projectId, clusterId)
+	cfg := api.EndpointCfg{Url: url, Method: http.MethodGet, SuccessStatus: http.StatusOK}
+	response, err := data.Client.Execute(
+		cfg,
+		nil,
+		data.Token,
+		nil,
+	)
+	if err != nil {
+		return nil, err
+	}
+	if err != nil {
+		return nil, err
+	}
+	clusterResp := clusterapi.GetClusterResponse{}
+	err = json.Unmarshal(response.Body, &clusterResp)
+	if err != nil {
+		return nil, err
+	}
+	clusterResp.Etag = response.Response.Header.Get("ETag")
+	return &clusterResp, nil
+}
+
+// This function takes a resource reference string and returns a resource.TestCheckFunc. The returned function, when used
+// in Terraform acceptance tests, ensures that the specified cluster resource exists in the Terraform state. It retrieves
+// the resource by name from the Terraform state and checks its existence. If the resource exists, it returns nil; otherwise,
+// it returns an error.
+func testAccExistsClusterResource(resourceReference string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		// retrieve the resource by name from state
+
+		var rawState map[string]string
+		for _, m := range s.Modules {
+			if len(m.Resources) > 0 {
+				if v, ok := m.Resources[resourceReference]; ok {
+					rawState = v.Primary.Attributes
+				}
+			}
+		}
+		fmt.Printf("raw state %s", rawState)
+		data, err := acctest.TestClient()
+		if err != nil {
+			return err
+		}
+		_, err = retrieveClusterFromServer(data, rawState["organization_id"], rawState["project_id"], rawState["id"])
+		if err != nil {
+			return err
+		}
+		return nil
 	}
 }
